Get rid of workers dir
